pyproject.toml
[tool.poetry]
name = "deneb"
version = "1.1.6"
description = "deneb project"
authors = ["DTailor <dann.croitoru@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
spotipy = "^2.9.0"
logzero = "^1.5.0"
click = "^7.0"
sentry-sdk = "^0.14.2"
tortoise-orm = "^0.15.18"
asyncpg = "^0.20.1"
aiohttp = "^3.6.2"
arrow = "^0.15.5"
pytz = "^2019.3"
sqlalchemy = "^1.3.13"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.2"
ipython = "^7.13.0"
isort = "^4.3.21"
mypy = "^0.761"
vulture = "^1.3"
fabric = "^2.5.0"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
pytest-asyncio = "^0.10.0"
aiomock = "^0.1.0"
coverage = "^5.0.3"
pylama = "^7.7.1"
black = "^19.10b0"
alembic = "^1.4.1"
psycopg2-binary = "^2.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
